♻️ refactor: move server modules (#3291)
