+++ src/server/routers/lambda/importer.ts
@@ -2,12 +2,21 @@ import { TRPCError } from '@trpc/server';
 import { z } from 'zod';
 
 import { authedProcedure, router } from '@/libs/trpc';
-import { S3 } from '@/server/files/s3';
-import { DataImporter } from '@/server/modules/DataImporter';
+import { S3 } from '@/server/modules/S3';
+import { DataImporterService } from '@/server/services/dataImporter';
 import { ImportResults, ImporterEntryData } from '@/types/importer';
 
+const importProcedure = authedProcedure.use(async (opts) => {
+  const { ctx } = opts;
+  const dataImporterService = new DataImporterService(ctx.userId);
+
+  return opts.next({
+    ctx: { dataImporterService },
+  });
+});
+
 export const importerRouter = router({
-  importByFile: authedProcedure
+  importByFile: importProcedure
     .input(z.object({ pathname: z.string() }))
     .mutation(async ({ input, ctx }): Promise<ImportResults> => {
       let data: ImporterEntryData | undefined;
@@ -27,12 +36,10 @@ export const importerRouter = router({
         });
       }
 
-      const dataImporter = new DataImporter(ctx.userId);
-
-      return dataImporter.importData(data);
+      return ctx.dataImporterService.importData(data);
     }),
 
-  importByPost: authedProcedure
+  importByPost: importProcedure
     .input(
       z.object({
         data: z.object({
@@ -45,8 +52,6 @@ export const importerRouter = router({
       }),
     )
     .mutation(async ({ input, ctx }): Promise<ImportResults> => {
-      const dataImporter = new DataImporter(ctx.userId);
-
-      return dataImporter.importData(input.data);
+      return ctx.dataImporterService.importData(input.data);
     }),
 });
